renames invalidation handler
